Добавляет исправления в модуль main и form
